use Word8 for Color8Code and Color8

+++ System/Console/ANSI/Codes.hs
@@ -138,7 +138,7 @@ sgrToCode sgr = case sgr of
   SetColor Foreground Vivid color -> [90 + colorToCode color]
   SetColor Background Dull color  -> [40 + colorToCode color]
   SetColor Background Vivid color -> [100 + colorToCode color]
-  SetColor8Code layer (Color8Code code) -> [38 + layerOffset layer, 5, code]
+  SetColor8Code layer (Color8Code code) -> [38 + layerOffset layer, 5, fromIntegral code]
   SetColor8 layer color -> sgrToCode $ SetColor8Code layer $ color8ToCode color
   SetRGBColor Foreground color -> [38, 2] ++ toRGB color
   SetRGBColor Background color -> [48, 2] ++ toRGB color
